% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{sumNum}
\alias{sumNum}
\title{Sums two numbers}
\usage{
sumNum(x, y)
}
\arguments{
\item{x, y}{The two numbers.}
}
\value{
The sum of the two numbers.
}
\description{
This function sums tow input numbers.
}
\examples{
sumNum(1,2)
}
